PAE-1790 Consolidate the public register into a self-contained module
The public register was spread across five separate trees: its repository port under domain, its logic under application, its S3 and in-memory adapters under adapters, its route under routes/v1 and its auditing under auditing. It now lives at src/public-register with application, repository and route layers and auditing alongside them, matching the shape the other feature modules already use.

Everything outside the module reaches it through src/public-register/index.js, which exposes the generate route and the two repository plugins and nothing else.

Generating the compliance report is now exposed on the reports module's public interface, so the public register no longer reaches into reports' internals to build it.
