+++ src/public-register/application/public-register-transformer.js
