+++ src/public-register/application/public-register-transformer.test.js
