+++ src/public-register/auditing.js
@@ -1,4 +1,8 @@
-import { extractUserDetails, recordSystemLog, safeAudit } from './helpers.js'
+import {
+  extractUserDetails,
+  recordSystemLog,
+  safeAudit
+} from '#root/auditing/helpers.js'
 
 /**
  * @import {SystemLogsRepository} from '#repositories/system-logs/port.js'
