+++ src/reports/index.js
@@ -6,6 +6,9 @@
  * @module reports
  */
 
+// Application exports
+export { generateReportCompliance } from './application/report-compliance.js'
+
 // Route exports
 export { reportsGet } from './routes/get.js'
 export { reportsGetDetail } from './routes/get-detail.js'
