--- src/routes/v1/public-register/index.js
@@ -1,1 +1,0 @@
-export { generateLatestPublicRegister } from './generate/post.js'
