+++ src/server/server.js
@@ -7,7 +7,7 @@ import HapiSwagger from 'hapi-swagger'
 import { secureContext } from '@defra/hapi-secure-context'
 
 import { dlqAdminPlugin } from '#plugins/dlq-admin.js'
-import { s3PublicRegisterRepositoryPlugin } from '#adapters/repositories/public-register/s3.plugin.js'
+import { s3PublicRegisterRepositoryPlugin } from '#public-register/index.js'
 import { s3UploadsRepositoryPlugin } from '#adapters/repositories/uploads/s3.plugin.js'
 import { sqsCommandExecutorPlugin } from '#adapters/sqs-command-executor/sqs-command-executor.plugin.js'
 import { failAction } from '#common/helpers/fail-action.js'
